+++ asserv/commandManager/CommandManager.h
@@ -14,11 +14,11 @@ public:
 	CommandManager(int capacity, ConsignController *ctrlr, Odometrie *odo);
 	~CommandManager();
 
-	bool addStraightLine(int64_t valueInmm);
-	bool addTurn(int64_t angleInDeg);
-	bool addGoTo(int64_t posXInmm, int64_t posYInmm);
-	bool addGoToEnchainement(int64_t posXInmm, int64_t posYInmm);
-	bool addGoToAngle(int64_t posXInmm, int64_t posYInmm);
+	bool addStraightLine(double valueInmm);
+	bool addTurn(double angleInDeg);
+	bool addGoTo(double posXInmm, double posYInmm);
+	bool addGoToEnchainement(double posXInmm, double posYInmm);
+	bool addGoToAngle(double posXInmm, double posYInmm);
 	void perform();
 
 	// Gestion d'un éventuel arrêt d'urgence
